Simplify end-of-animation-detection example
